Modularize the input_data package structure
