+++ src/input_data/structure/__init__.py
@@ -1,0 +1,7 @@
+"""A package describing the basic structure of datasets, including abstract base classes and enums."""
+from .base import ManagedDataset, DatasetInfo
+
+__all__ = [
+    "ManagedDataset",
+    "DatasetInfo",
+]
